class declaration to header
