+++ src/rvoice/fluid_rev_freeverb.cpp
@@ -9,11 +9,7 @@
   Translated to C by Peter Hanappe, Mai 2001
 */
 
-#include "fluid_sys.h"
-#include "fluid_rev.h"
-#include "fluid_rev_filters.h"
-
-#include <new>
+#include "fluid_rev_freeverb.h"
 
 /***************************************************************
  *
@@ -39,11 +35,6 @@
  */
 #define DC_OFFSET ((fluid_real_t)1e-8)
 
-using fluid_allpass = fluid_reverb_allpass<fluid_real_t>;
-using fluid_comb = fluid_reverb_comb<fluid_real_t>;
-
-#define numcombs 8
-#define numallpasses 4
 #define	fixedgain 0.015f
 /* scale_wet_width is a compensation weight factor to get an output
    amplitude (wet) rather independent of the width setting.
@@ -89,40 +80,6 @@ using fluid_comb = fluid_reverb_comb<fluid_real_t>;
 #define allpasstuningR3 (341 + stereospread)
 #define allpasstuningL4 225
 #define allpasstuningR4 (225 + stereospread)
-
-struct fluid_revmodel_freeverb : public _fluid_revmodel_t
-{
-    fluid_real_t roomsize;
-    fluid_real_t damp;
-    fluid_real_t level, wet1, wet2;
-    fluid_real_t width;
-    fluid_real_t gain;
-    /*
-     The following are all declared inline
-     to remove the need for dynamic allocation
-     with its subsequent error-checking messiness
-    */
-    /* Comb filters */
-    fluid_comb combL[numcombs];
-    fluid_comb combR[numcombs];
-    /* Allpass filters */
-    fluid_allpass allpassL[numallpasses];
-    fluid_allpass allpassR[numallpasses];
-
-    explicit fluid_revmodel_freeverb(fluid_real_t sample_rate);
-    ~fluid_revmodel_freeverb() override;
-
-    void processmix(const fluid_real_t *in, fluid_real_t *left_out,
-                    fluid_real_t *right_out) override;
-    void processreplace(const fluid_real_t *in, fluid_real_t *left_out,
-                        fluid_real_t *right_out) override;
-    void reset() override;
-    void set(int set, fluid_real_t roomsize, fluid_real_t damping,
-             fluid_real_t width, fluid_real_t level) override;
-    int samplerate_change(fluid_real_t sample_rate) override;
-};
-
-typedef struct fluid_revmodel_freeverb fluid_revmodel_freeverb_t;
 
 static void fluid_freeverb_revmodel_update(fluid_revmodel_freeverb_t *rev);
 static void fluid_freeverb_revmodel_init(fluid_revmodel_freeverb_t *rev);
